TCP input: remove usage of static variables
This is being done as part of the changes needed to allow
concurrent reception of multiple TCP input packets.
Static variables are now replaced by local variables declared in
tcp_input() and passed to various functions involved in processing
of input packets. Removal of the tcp_input_pcb variable creates a
potential issue if a TCP PCB is deallocated during input packet
processing; this issue will be solved in the next commit by adding
reference counting to the TCP PCB structure.