Raw API: implement multi-thread safety
This change adds locking and reference counting mechanisms to allow
lwIP raw API functions and input callbacks to be invoked
concurrently from multiple threads.
A new type (sys_lock_t) and a set of macros (such as SYS_ARCH_LOCK
and SYS_ARCH_UNLOCK) are being added so that applications can
provide a platform-specific locking implementation. Various
subsystems in the network stack (DNS, IP, DHCP, ARP, MLD, ND,
netif, TCP, UDP) now use locking to protect their global data
structures from concurrent access.
The SYS_ARCH_INC() and SYS_ARCH_DEC() macros are used to atomically
increment and decrement a variable, respectively.
The tcp_pcb structure is locked in order to protect from concurrent
access the data representing a TCP connection state. In order to
avoid deadlock, the lock of a TCP PCB must not be acquired while
holding the global TCP lock (unless tcp_trylock() is used).
Application callbacks registered on a TCP PCB are invoked with the
PCB lock held; the application must acquire the PCB lock (via the
tcp_lock() macro) before calling raw API functions (such as
tcp_connect or tcp_output) on a TCP PCB.
The ip_pcb structure is reference-counted so that it is possible
to process PCBs retrieved from global lists (e.g. the active TCP
PCB list) without holding the lock that protects the lists, while
at the same time allowing the application code to invoke APIs
(such as tcp_close ot udp_remove) to deregister or deallocate a
PCB.
Network interfaces are reference-counted so that it is possible to
remove an interface while network packets are output on it; the
netif_ref() and netif_unref() macros are provided to increment and
decrement the interface reference count, respectively.
The NETIF_FOREACH() macro is being replaced by a multithread-safe
netif_iterate() function that takes a callback as argument and
invokes the callback for each registered network interface.
The netif_poll_all() function is being removed because a network
interface cannot be polled while holding the netif lock (IP input
parsing code invokes netif_iterate(), which needs to acquire the
lock); the netif_poll_loopback() is now defined to poll the
loopback interface, which is the only interface that needs to be
polled.
The netif_get_default() function is now used to retrieve the
default network interface instead of accessing directly the
netif_default global variable.