Eliminate deadlocks when calling tcp_pcbs_sane()
The tcp_pcbs_sane() function was locking the tcp_mutex in order to
safely traverse the pcbs list, but it is sometimes called from
other places such as tcp_pcb_remove(), which is called with that
lock already held, thus creating a deadlock.

This change removes the locking from tcp_pcbs_sane and adds it
to the one place where it is called without the lock held.