TCP: add support for switching state from TIME-WAIT to ACTIVE
RFC 1122 (section 4.2.2.13) says: "A connection in TIME-WAIT state
may accept a new SYN from the remote TCP to reopen the connection
directly from TIME-WAIT state, if it assigns its initial sequence
number for the new connection to be larger than the largest
sequence number it used on the previous connection incarnation".
This change amends the TCP input handling code to support accepting
new connections for sockets in TIME-WAIT state.