Cyclic timers: dynamically enable and disable MLD6 timer
Similarly to the TCP timer, the MLD6 timer does not always need to
be dispatched. Dispatching an unneeded timer causes useless CPU
usage, especially for high-frequency timers (the MLD6 timer period
is 100 ms).
In order to decrease CPU usage when the system is idle, this change
removes the MLD6 timer from the list of static timers and adds
dynamic enablement of the timer.