IPv4: add support for setting DONT_FRAGMENT flag in output packets
Add a new `pmtudisc` field to the ip_pcb struct to store the path
MTU discovery setting for a socket; if the value of this field is
IP_PMTUDISC_DO or IP_PMTUDISC_PROBE, the DF flag is set in output
IPv4 packets (if a packet is larger than the MTU of the output
interface, the packet is not sent and ERR_MSGSIZE is returned); the
default value of this field is IP_PMTUDISC_DONT, i.e. the DF flag
is not set by default.